Collisions refactored
